qdrant_init/pyproject.toml
[tool.poetry]
name = "cluster-preparation"
version = "0.1.0"
description = ""
authors = ["FatMadLad <mrammler@web.de>"]
packages = [{include = "cluster_preparation"}]

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
numpy = "^1.24.2"
sentence-transformers = "^2.2.2"
pillow = "^9.5.0"
qdrant-client = "1.1.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
